Update index.html.template
I have had the Windows element app update two times over the last year.  Both time after updating, the --proxy-server=socks5://127.0.0.1:9050 gets removed from the shortcut making element unable to communicate via tor.  In addition, the update add a new path under the "start in box" is added pointing to the version of element to run.  That needs to be deleted.  Once these two steps are completed and the Element app is restarted the connection resumes.